Suggestion on how to handle phone-numbers
Contact phones are hairy. Most brokers want both a personal and "daytime contact". However some does not, forcing any model does seem a bit shoehordned and in practice the general approach will be to first call the cellphone and then work your way through the list.
Apparently the example-employee is located in Töreboda-Hova so i have updated the personal-phone to reflect this